+++ core/src/main/java/org/bookwoori/core/domain/climbingMember/repository/ClimbingMemberRepository.java
@@ -1,0 +1,8 @@
+package org.bookwoori.core.domain.climbingMember.repository;
+
+import org.bookwoori.core.domain.climbingMember.entity.ClimbingMember;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ClimbingMemberRepository extends JpaRepository<ClimbingMember, Long> {
+
+}
