--- core/src/main/java/org/bookwoori/core/domain/member/entity/ServerRole.java
@@ -1,6 +1,0 @@
-package org.bookwoori.core.domain.member.entity;
-
-public enum ServerRole {
-    OWNER,
-    MEMBER;
-}
