+++ core/src/main/java/org/bookwoori/core/domain/serverMember/entity/ServerMember.java
@@ -1,7 +1,9 @@
-package org.bookwoori.core.domain.member.entity;
+package org.bookwoori.core.domain.serverMember.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
 import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
@@ -13,33 +15,33 @@ import jakarta.validation.constraints.NotNull;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import org.bookwoori.core.domain.climbing.entity.Climbing;
+import org.bookwoori.core.domain.member.entity.Member;
+import org.bookwoori.core.domain.server.entity.Server;
 
 @Entity
-@Table(name = "climbing_member")
+@Table(name = "server_member")
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class ClimbingMember {
+public class ServerMember {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "climbing_member_id", updatable = false)
-    private Long climbingMemberId;
+    @Column(name = "server_member_id", updatable = false)
+    private Long serverMemberId;
 
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "climbing_id", updatable = false)
-    @NotNull
-    private Climbing climbing;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "member_id", updatable = false)
+    @JoinColumn(name = "member_id")
     @NotNull
     private Member member;
 
-    @Column(name = "has_shared")
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "server_id")
     @NotNull
-    private boolean hasShared;
+    private Server server;
 
-    @Column(name = "memo")
-    private String memo;
+    @Column(name = "role")
+    @NotNull
+    @Enumerated(EnumType.STRING)
+    private ServerRole role;
+
 }
