+++ core/src/main/java/org/bookwoori/core/domain/serverMember/entity/ServerRole.java
@@ -1,0 +1,6 @@
+package org.bookwoori.core.domain.serverMember.entity;
+
+public enum ServerRole {
+    OWNER,
+    MEMBER;
+}
