+++ core/src/main/java/org/bookwoori/core/domain/serverMember/repository/ServerMemberRepository.java
@@ -1,0 +1,8 @@
+package org.bookwoori.core.domain.serverMember.repository;
+
+import org.bookwoori.core.domain.serverMember.entity.ServerMember;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ServerMemberRepository extends JpaRepository<ServerMember, Long> {
+
+}
